+++ demo/app.module.ts
@@ -5,8 +5,10 @@ import { FormsModule } from '@angular/forms';
 import { APP_BASE_HREF, Location } from '@angular/common';
 import { AppComponent } from './app.component';
 import { SparklineComponent } from './sparkline/sparkline.component';
+import { TimelineFilterBarChartComponent } from './timeline-filter-bar-chart/timeline-filter-bar-chart.component';
 import { NgxChartsModule } from '../src';
 import { NgxUIModule } from '@swimlane/ngx-ui';
+import { ComboChartComponent, ComboSeriesVerticalComponent } from './combo-chart';
 
 @NgModule({
   providers: [
@@ -15,8 +17,20 @@ import { NgxUIModule } from '@swimlane/ngx-ui';
       useFactory: getBaseLocation
     }
   ],
-  imports: [NgxChartsModule, BrowserModule, BrowserAnimationsModule, FormsModule, NgxUIModule],
-  declarations: [AppComponent, SparklineComponent],
+  imports: [
+    NgxChartsModule,
+    BrowserModule,
+    BrowserAnimationsModule,
+    FormsModule,
+    NgxUIModule
+  ],
+  declarations: [
+    AppComponent,
+    SparklineComponent,
+    TimelineFilterBarChartComponent,
+    ComboChartComponent,
+    ComboSeriesVerticalComponent
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }