+++ src/common/axes/y-axis.component.ts
@@ -25,9 +25,13 @@ import { YAxisTicksComponent } from './y-axis-ticks.component';
         [orient]="yOrient"
         [showGridLines]="showGridLines"
         [gridLineWidth]="dims.width"
+        [referenceLines]="referenceLines"
+        [showRefLines]="showRefLines"
+        [showRefLabels]="showRefLabels"
         [height]="dims.height"
         (dimensionsChanged)="emitTicksWidth($event)"
       />
+
       <svg:g ngx-charts-axis-label
         *ngIf="showLabel"
         [label]="labelText"
@@ -51,7 +55,9 @@ export class YAxisComponent implements OnChanges {
   @Input() yAxisTickInterval;
   @Input() yAxisTickCount: any;
   @Input() yOrient: string = 'left';
-  
+  @Input() referenceLines;
+  @Input() showRefLines;
+  @Input() showRefLabels;
   @Output() dimensionsChanged = new EventEmitter();
 
   yAxisClassName: string = 'y axis';
@@ -59,8 +65,7 @@ export class YAxisComponent implements OnChanges {
   offset: any;
   transform: any;
   yAxisOffset: number = -5;
-
-  labelOffset: number = 80;
+  labelOffset: number = 15;
   fill: string = 'none';
   stroke: string = '#CCC';
   tickStroke: string = '#CCC';
@@ -75,6 +80,7 @@ export class YAxisComponent implements OnChanges {
   update(): void {
     this.offset = this.yAxisOffset;
     if (this.yOrient === 'right') {
+      this.labelOffset = 65;
       this.transform = `translate(${this.offset + this.dims.width} , 0)`;
     } else {
       this.transform = `translate(${this.offset} , 0)`;
@@ -86,7 +92,12 @@ export class YAxisComponent implements OnChanges {
   }
 
   emitTicksWidth({ width }): void {
-    if (width !== this.labelOffset) {
+    if (width !== this.labelOffset && this.yOrient === 'right' ) {
+      this.labelOffset = width + this.labelOffset;
+      setTimeout(() => {
+        this.dimensionsChanged.emit({width});
+      }, 0);
+    } else if (width !== this.labelOffset) {
       this.labelOffset = width;
       setTimeout(() => {
         this.dimensionsChanged.emit({width});